kexec-lite: add zImage support
The zimage boot wrapper files contain a ".kernel:vmlinux.strip"
elf section in which is compressed a vmlinux file. The patch adds
support for zimage files by simply unzipping the content of this
section in a tmp file. This file is then loaded as the standard
vmlinux file.

The code is a bit brutal and duplicates some elf parts which could
be made common.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>